Fix type conversion, long provided, int expected

+++ app/src/main/java/com/snow/dreamdiary/feature_dreams/presentation/dailysurvey/DailySurveyViewModel.kt
@@ -104,7 +104,7 @@ class DailySurveyViewModel @Inject constructor(
 
                     _state.value = state.value.copy(
                         surveyData = state.value.surveyData.copy(
-                            timeSlept = timesSlept.toLong(),
+                            timeSlept = timesSlept,
                             dreamsNum = if (
                                 state.value.surveyData.didDream
                             )
